+++ R/shoji-folder.R
@@ -86,7 +86,8 @@ parentFolderURL <- function(x) {
     }
 }
 
-rootFolder <- function(x) {
+walkFoldersToRoot <- function(x) {
+    ## Walk the path up through the parents until you can go no farther
     this <- folder(x)
     ## If the parent of x is NULL, we're already at top level.
     while (!is.null(this)) {
@@ -95,6 +96,8 @@ rootFolder <- function(x) {
     }
     return(x)
 }
+
+setMethod("rootFolder", "ShojiFolder", walkFoldersToRoot)
 
 createFolder <- function(where, name, index, ...) {
     ## TODO: include index of variables/folders in a single request;
@@ -116,42 +119,6 @@ createFolder <- function(where, name, index, ...) {
 setMethod("name<-", "ShojiFolder",
     function(x, value) setEntitySlot(x, "name", value)
 )
-
-#' @rdname delete
-#' @export
-setMethod("delete", "ShojiFolder", function(x, ...) {
-    if (is.null(parentFolderURL(x))) {
-        halt("Cannot delete root folder")
-    }
-
-    # count the variable/folder objects, and warn the user that they will be
-    # summarily deleted as well. Projects must be empty to be deleted (which is
-    # enforced on the server, so we only need to check VariableFolders) send as
-    # a message before the prompt for test-ability, and so the prompt isn't lost
-    # at then end of a long line.
-    if (inherits(x, "VariableFolder")) {
-        obj_names <- names(x)
-        num_vars <- length(x)
-        obj_word <- ifelse(num_vars > 1, "objects", "object")
-
-        if (num_vars > 5) {
-            obj_string <- serialPaste(dQuote(head(obj_names, 5)), "...")
-        } else {
-            obj_string <- serialPaste(dQuote(obj_names))
-        }
-        message(
-            "This folder contains ", num_vars, " ", obj_word, ": ", obj_string, 
-            ". Deleting the folder will also delete these objects (including ",
-            "their contents)."
-        )
-    }
-
-    if (!askForPermission(paste0("Really delete ", name(x), "?"))) {
-        halt("Must confirm deleting folder")
-    }
-    out <- crDELETE(self(x))
-    invisible(out)
-})
 
 #' Change the order of entities in folder
 #'