+++ R/tuple.R
@@ -91,22 +91,6 @@ setMethod("id", "DatasetTuple", function(x) x$id)
 
 #' @rdname tuple-methods
 #' @export
-setMethod("delete", "ShojiTuple", function(x, ...) {
-    crDELETE(x@entity_url, drop = dropCache(x@index_url))
-})
-#' @rdname tuple-methods
-#' @export
-setMethod("delete", "DatasetTuple", function(x, ...) {
-    prompt <- paste0("Really delete dataset ", dQuote(name(x)), "?")
-    if (!askForPermission(prompt)) {
-        halt("Must confirm deleting dataset")
-    }
-    out <- callNextMethod()
-    invisible(out)
-})
-
-#' @rdname tuple-methods
-#' @export
 setMethod("name", "ShojiTuple", vget("name"))
 
 #' @rdname tuple-methods