+++ tests/testthat/test-variable-delete.R
@@ -1,5 +1,10 @@
 context("Deleting variables")
 
+test_that("delete() method fallback", {
+    expect_error(delete(NULL), "'delete' only valid for Crunch objects")
+    expect_error(delete(list()), "'delete' only valid for Crunch objects")
+})
+
 with_mock_crunch({
     ds <- loadDataset("test ds")
     test_that("Assigning NULL doesn't ask you about deleting 0 variables", {